+++ js/main.js
@@ -2,111 +2,162 @@
 //jshint -W097
 'use strict';
 
-// require('../lib/jquery-2.2.0.min.js')
-// require('../lib/two.min.js')
-require('./keybindings.js')
-// require('./util.js')
+import * as d3 from 'd3'
+import { getCookie, setCookie } from './util.js';
 
-import * as d3 from "d3"
-import * as utils from './util.js';
+const NOCOLOR  = 'rgba(0, 0, 0, 0)'
+
+// const GREYBLUE = 'rgb(54, 60, 77)'
+// const ORANGE   = 'rgb(228, 142, 13)'
+// const AQUA     = 'rgb(81, 191, 222)'
+
+const GREY     = '#E8E8EA'
+const YELLOW   = '#F0C034'
+const GREEN    = '#54BA75'
+const RED      = '#C83D34'
+const DARKBLUE = '#234B6E'
 
 class Game {
   constructor() {
     this.isMobile = typeof window.orientation !== 'undefined'
     this.running = true
-
-    // Make an instance of two and place it on the page.
-    this.elem = document.getElementById('draw-shapes')
-    this.params = { width: window.innerWidth, height: window.innerHeight }
-    this.two = new Two(this.params).appendTo(this.elem)
-
-    this.centerRadius = this.two.height / 8
-    this.centerX = this.two.width / 2
-    this.centerY = this.two.height / 2
-    this.thickness = this.two.height / 20
-
-    this.background = this.two.makeRectangle(this.centerX, this.centerY, this.two.width, this.two.height)
-    this.centerCircle = this.two.makeCircle(this.centerX, this.centerY,  this.centerRadius)
-    this.smallCircle = this.two.makeEllipse(
-      this.centerX - this.centerRadius, this.centerY,
-      this.thickness * 0.4, this.thickness * 0.4
-    )
-    this.tick = this.two.makeRoundedRectangle(
-      this.centerX - this.centerRadius, this.centerY,
-      this.thickness / 2, this.thickness * 1.9, 3
-    )
-
-    this.background.fill = 'rgb(109, 140, 149)'
-    this.background.noStroke()
-
-    this.centerCircle.noFill()
-    this.centerCircle.stroke = 'rgb(54, 60, 77)'
-    this.centerCircle.linewidth = this.thickness
-
-    this.smallCircle.fill = 'rgb(228, 142, 13)'
-    this.smallCircle.noStroke()
-
-    this.tick.rotation = Math.PI / 2
-    this.tick.noStroke()
-    this.tick.fill = 'rgb(81, 191, 222)'
-
-    this.highScore = parseFloat(this.getCookie('highScore2')) || 0
+    this.highScore = parseFloat(getCookie('highScore2')) || 0
     this.tickAngle = 0
-    this.points = 0
-    this.randAngle = 2 * Math.PI * Math.random()
+    this.targetAngle = 360 * Math.random()
+    this._points = 0
     this.direction = 1
+    this._misses = 0
+    this.willMiss = false
+    this.colorQueue = []
+    this.isTransitioning = false
+
+    this.container = d3.select('body')
+      .append('svg')
+      .attr('id', 'draw-shapes')
+      .attr('width', window.innerWidth)
+      .attr('height', window.innerHeight)
+      .style('background-color', GREY)
+
+    this.shapeContainer = this.container
+      .append('g')
+      .attr('transform', `translate(${ this.centerX }, ${ this.centerY })`)
+
+    this.centerShape = this.shapeContainer
+      .append('circle')
+      .attr('r', this.centerRadius)
+      .attr('fill', NOCOLOR)
+      .style('stroke', GREEN)
+      .style('stroke-width', this.thickness)
+
+    this.targetCircle = this.shapeContainer
+      .append('circle')
+      .attr('cy', this.centerRadius)
+      .attr('r', this.thickness / 2.5)
+      .attr('fill', GREY)
+      .attr('transition-timing-function', 'ease-in-out')
+
+    this.tick = this.shapeContainer
+      .append('rect')
+      .attr('x', -this.centerRadius / 20)
+      .attr('y', this.centerRadius - this.thickness * 1.2)
+      .attr('width', this.centerRadius / 10)
+      .attr('height', this.thickness * 2.4)
+      .attr('fill', DARKBLUE)
+      .attr('rx', 5)
+      .attr('ry', 5)
+      .attr('transition-timing-function', 'linear')
 
     this.updateScore()
+    this.moveTargetCircle()
+    this.updateGame()
 
-    this.two.bind('update', (frameCount) => {
-      if (this.running) {
-        this.updateGame()
+    window.addEventListener('keydown', (event) => {
+      switch (event.which) {
+        case 32: // space key
+          this.onSpaceBar(); break
+        case 27: // esc
+          this.running = !this.running; break
       }
-    }).play()
-
-    let ten = (10 * (Math.PI / 180)) // ten degrees in radians
-
-    $(window).on('key-esc', () => this.running = !this.running )
-
-    $(window).on('key-space', () => {
-      if (this.tick.rotation > this.smallCircle.rotation - ten &&
-          this.tick.rotation < this.smallCircle.rotation + ten) {
-        this.direction *= -1
-        this.randAngle += Math.PI * Math.random()
-        this.randAngle %= 2 * Math.PI
-        this.points += 1
-      } else {
-        this.points = 0
-      }
-
-      this.updateScore()
     })
   }
 
+  onSpaceBar() {
+    if (this.tickOverLapsCircle()) {
+      this.points++
+      this.willMiss = false
+      this.misses = 0
+      this.direction *= -1
+      this.targetAngle += (90 + 180 * Math.random()) * -this.direction
+      if (this.targetAngle < 0)
+        this.targetAngle = 360 + this.targetAngle
+      this.targetAngle %= 360
+      this.moveTargetCircle()
+    } else {
+      this.misses = 0
+      this.points = 0
+    }
+
+    this.updateScore()
+  }
+
+  moveTargetCircle() {
+    this.targetCircle
+      .transition()
+      .attr('transform', `rotate(${ this.targetAngle })`)
+      .duration(500)
+  }
+
+  tickOverLapsCircle() {
+    return (
+      Math.abs(this.tickAngle) > this.targetAngle - 10 &&
+      Math.abs(this.tickAngle) < this.targetAngle + 10
+    )
+  }
+
   updateGame() {
-    let tickSpeed = this.points * 0.001
-    this.tickAngle += (0.05 + tickSpeed) * this.direction; this.tickAngle %= 2 * Math.PI
+    setTimeout(this.updateGame.bind(this), 16)
 
-    this.tick.translation.x = this.centerX + (this.centerRadius * Math.cos(this.tickAngle))
-    this.tick.translation.y = this.centerY + (this.centerRadius * Math.sin(this.tickAngle))
-    this.tick.rotation = (Math.PI / 2) + Math.atan2(
-      Math.sin(this.tickAngle),
-      Math.cos(this.tickAngle)
-    )
+    if ( this.running ) {
+      let tickSpeed = (3 + this.points * 0.1) * this.direction
+      if (this.tickAngle + tickSpeed > 0){
+        this.tickAngle += tickSpeed
+      } else {
+        this.tickAngle = 360 + tickSpeed
+      }
 
-    this.smallCircle.translation.x = this.centerX + (this.centerRadius * Math.cos(this.randAngle))
-    this.smallCircle.translation.y = this.centerY + (this.centerRadius * Math.sin(this.randAngle))
-    this.smallCircle.rotation = (Math.PI / 2) + Math.atan2(
-      Math.sin(this.randAngle),
-      Math.cos(this.randAngle)
-    )
+      this.tickAngle %= 360
+
+      if (this.tickOverLapsCircle()) this.willMiss = true
+
+      if (this.willMiss && !this.tickOverLapsCircle()) {
+        this.misses++
+        this.willMiss = false
+      }
+
+      if (this.colorQueue.length > 0 && !this.isTransitioning) {
+        this.isTransitioning = true
+        this.centerShape
+          .transition()
+          .style('stroke', this.colorQueue[0])
+          .duration(300)
+          .on('end', () => {
+            this.isTransitioning = false
+            this.colorQueue.splice(0, 1)
+          })
+      }
+
+      this.tick
+        .transition()
+        .attr('transform', `rotate(${ this.tickAngle })`)
+        .duration(15)
+    }
   }
 
   updateScore() {
     this.highScore = (this.highScore > this.points ? this.highScore : this.points)
     document.getElementById('high-score').innerHTML = 'High Score: ' + this.highScore
     document.getElementById('score').innerHTML = this.points
-    this.setCookie('highScore2', this.highScore, 9999999999999)
+    setCookie('highScore2', this.highScore, 9999999999999)
     let instructions = document.getElementById('instructions')
     if (this.points === 0)
       instructions.innerHTML = `${ this.isMobile ? 'tap' : 'press the spacebar'} to begin`
@@ -114,22 +165,56 @@ class Game {
       instructions.innerHTML = ''
   }
 
-  setCookie(cname, cvalue, exdays) {
-    var d = new Date();
-    d.setTime(d.getTime() + (exdays*24*60*60*1000));
-    var expires = "expires="+d.toUTCString();
-    document.cookie = cname + "=" + cvalue + "; " + expires;
+  get centerRadius() {
+    return window.innerHeight < window.innerWidth
+      ? window.innerHeight / 6
+      : window.innerWidth  / 6
+  }
+  get centerX() { return window.innerWidth / 2 }
+  get centerY() { return window.innerHeight / 2 }
+  get thickness() { return this.centerRadius / 2.5 }
+
+  set points(val) {
+    this._points = val
   }
 
-  getCookie(cname) {
-    var name = cname + "=";
-    var ca = document.cookie.split(';');
-    for(var i = 0; i<ca.length; i++) {
-        var c = ca[i];
-        while (c.charAt(0) === ' ') c = c.substring(1);
-        if (c.indexOf(name) === 0) return c.substring(name.length,c.length);
+  get points() {
+    return this._points
+  }
+
+  set centerShapeColor(val) {
+    this.colorQueue.push(val)
+  }
+
+  set misses(val) {
+    if (this.points === 0) {
+      this.centerShapeColor = GREEN
+      return
     }
-    return "";
+
+    this._misses = val
+
+    if (this._misses === 0)
+      this.centerShapeColor = GREEN
+
+    else if (this._misses === 1)
+      this.centerShapeColor = YELLOW
+
+    else if (this._misses === 2)
+      this.centerShapeColor = RED
+
+    else if (this._misses >= 3) {
+      if (this.points - 1 < 0)
+        this.centerShapeColor = GREEN
+      else
+        this.points--
+
+      this.updateScore()
+    }
+  }
+
+  get misses() {
+    return this._misses
   }
 }
